webapps/balsaFe/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
gunicorn==20.1.0
google-cloud-firestore==2.13.0
openai==1.2.3
python-dotenv==1.0.0